app/src/main/java/com/example/recipestask/modules/home/domain/interactors/GetRecipesUseCase.kt
package com.example.recipestask.modules.home.domain.interactors

import com.example.recipestask.modules.home.domain.entity.RecipesEntity
import com.example.recipestask.modules.home.domain.repository.RecipesRepository
import kotlinx.coroutines.CoroutineDispatcher
import kotlinx.coroutines.Dispatchers
import kotlinx.coroutines.flow.Flow
import kotlinx.coroutines.flow.flowOn
import javax.inject.Inject

class GetRecipesUseCase @Inject constructor(private val recipesRepository: RecipesRepository) {
    fun build(): Flow<List<RecipesEntity>> {
        return recipesRepository.getRecipes()
    }
}